+++ src/cc_malloc.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void* cc_malloc(int nbytes, int zero);
+void cc_free(void* m);
+void cc_free_all(void);